+++ Changelog.md
@@ -2,12 +2,18 @@
 
 ### Ignition Gazebo 2.X.X
 
+1. Support conversion and serialization of scene and light components
+    * [Pull Request 297](https://bitbucket.org/ignitionrobotics/ign-gazebo/pull-requests/297)
+
 1. Use scene ambient and background color information in sensor
-   configuration. 
+   configuration.
     * [Pull Request 268](https://bitbucket.org/ignitionrobotics/ign-gazebo/pull-requests/268)
 
 1. Added an SDF message to the start of log files.
     * [Pull Request 257](https://bitbucket.org/ignitionrobotics/ign-gazebo/pull-requests/257)
+    
+1. Update Magnetometer component to use sdf::Sensor object instead of an sdf::ElementPtr.
+    * [Pull Request 272](https://bitbucket.org/ignitionrobotics/ign-gazebo/pull-requests/272)
 
 ## Ignition Gazebo 1.x